requirements.txt
psutil>=5.9.0
pyyaml>=6.0.1
